third_party/blink/web_tests/platform/generic/http/tests/devtools/elements/styles-3/styles-disable-property-after-selector-edit-expected.txt

Verifies that sequence of setting selector and disabling property works.


Running: selectInitialNode

Running: editSelector

Running: testDisableProperty


#### AFTER PROPERTY DISABLED ####


[expanded] 
element.style { ()

[expanded] 
#inspected, .INSERTED-OTHER-SELECTOR { (<style>)
/-- overloaded --/ /-- disabled --/     /* color: red; */

[expanded] 
div { (user agent stylesheet)
    display: block;
